+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 #### 2019.4.0 - Feature release
+- Added Nextcloud 16 support
 - Improved custom fields
 - Fixed double sse key generation
 - Fixed tags being not removed from passwords